requirements.txt
streamlit
streamlit_option_menu
qrcode
pillow
python-barcode

